Acceptance: Parallelize `CSBS` tests
Make `CSBS` tests parallel

Make use of quotas in CSBS tests
