--- opentelekomcloud/acceptance/csbs/import_opentelekomcloud_csbs_backup_v1_test.go
@@ -1,27 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccCSBSBackupV1_importBasic(t *testing.T) {
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCSBSBackupV1Destroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCSBSBackupV1Basic,
-			},
-			{
-				ResourceName:      resourceBackupName,
-				ImportState:       true,
-				ImportStateVerify: false,
-			},
-		},
-	})
-}
